+++ Plugins/Events/Events.hpp
@@ -31,6 +31,7 @@ class InventoryEvents;
 class TrapEvents;
 class TimingBarEvents;
 class LevelEvents;
+class PVPEvents;
 
 class Events : public NWNXLib::Plugin
 {
@@ -110,6 +111,7 @@ private:
     std::unique_ptr<TrapEvents> m_trapEvents;
     std::unique_ptr<TimingBarEvents> m_timingBarEvents;
     std::unique_ptr<LevelEvents> m_levelEvents;
+    std::unique_ptr<PVPEvents> m_PVPEvents;
 };
 
 }