+++ Plugins/Tweaks/Tweaks.hpp
@@ -17,6 +17,7 @@ class HideDMsOnCharList;
 class DisableMonkAbilitiesWhenPolymorphed;
 class StringToIntBaseToAuto;
 class DeadCreatureFiresOnAreaExit;
+class PreserveActionsOnDMPossess;
 
 class Tweaks : public NWNXLib::Plugin
 {
@@ -37,6 +38,7 @@ private:
     std::unique_ptr<DisableMonkAbilitiesWhenPolymorphed> m_DisableMonkAbilitiesWhenPolymorphed;
     std::unique_ptr<StringToIntBaseToAuto> m_StringToIntBaseToAuto;
     std::unique_ptr<DeadCreatureFiresOnAreaExit> m_DeadCreatureFiresOnAreaExit;
+    std::unique_ptr<PreserveActionsOnDMPossess> m_PreserveActionsOnDMPossess;
 };
 
 }