+++ CHANGELOG.md
@@ -34,6 +34,7 @@ NOTICE: The ABI has changed, please make sure to update your nwnx.nss and recomp
 - Events: Added WebHook Success/Failure events with rate limit feedback
 - Events: Added UseLoreOnItem and PayToIdentifyItem events
 - Events: Added {Add|Remove}Gold events to InventoryEvents
+- Events: Added PVP Attitude Change event
 - Profiler: Support profiler perf scopes via nwscript
 - SQL: Added support for SQLite
 - Tweaks: DisableQuickSave
@@ -41,8 +42,10 @@ NOTICE: The ABI has changed, please make sure to update your nwnx.nss and recomp
 - Tweaks: DisableMonkAbilitiesWhenPolymorphed
 - Tweaks: StringToIntBaseToAuto
 - Tweaks: DeadCreatureFiresOnAreaExit
+- Tweaks: PreserveActionsOnDMPossess
 - Weapon: Feat and Base Item names were added to LOG_INFO feedback
 - WebHook: Added support for richer Slack-compatible messages
+- Util: Added the environment variable `NWNX_UTIL_PRE_MODULE_START_SCRIPT=scriptname` which lets you set a nwscript that runs before the OnModuleLoad event
 ##### New Plugins
 The following plugins were added:
 - **Appearance**: Allows the appearance and some other things of creatures to be overridden per player
@@ -137,6 +140,7 @@ The following plugins were added:
 - Object: SetPlaceableIsStatic()
 - Object: {Get|Set}AutoRemoveKey()
 - Object: {Get|Set}TriggerGeometry()
+- Object: {Add|Remove}IconEffect()
 - Player: ShowVisualEffect()
 - Player: ForcePlaceableInventoryWindow()
 - Player: MusicBackgroundChangeDay()